+++ code/packages/create-frontend-app/template/+nextjs+azure-web-pubsub/components/room-components/overlay/node-properties.tsx
@@ -3,12 +3,12 @@
 import React from 'react';
 import { useWeave } from '@inditextech/weave-react';
 import { useCollaborationRoom } from '@/store/store';
-import { PositionProperties } from './../node-properties/position-properties';
-import { SizeProperties } from './../node-properties/size-properties';
-import { AppearanceProperties } from './../node-properties/appearance-properties';
-import { FillProperties } from './../node-properties/fill-properties';
-import { StrokeProperties } from './../node-properties/stroke-properties';
-import { TextProperties } from './../node-properties/text-properties';
+import { PositionProperties } from '../node-properties/position-properties';
+import { SizeProperties } from '../node-properties/size-properties';
+import { AppearanceProperties } from '../node-properties/appearance-properties';
+import { FillProperties } from '../node-properties/fill-properties';
+import { StrokeProperties } from '../node-properties/stroke-properties';
+import { TextProperties } from '../node-properties/text-properties';
 import { ImageProperties } from '../node-properties/image-properties';
 import { ColorTokenProperties } from '../node-properties/color-token-properties';
 import { FrameProperties } from '../node-properties/frame-properties';
@@ -17,6 +17,7 @@ import { SIDEBAR_ELEMENTS } from '@/lib/constants';
 import { X } from 'lucide-react';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { WeaveSelection } from '@inditextech/weave-types';
+import { MetaProperties } from '../node-properties/meta-properties';
 
 export const NodeProperties = () => {
   const instance = useWeave((state) => state.instance);
@@ -194,6 +195,7 @@ export const NodeProperties = () => {
       </div>
       <ScrollArea className="w-full h-[calc(100%-95px)]">
         <div className="w-full flex flex-col">
+          <MetaProperties />
           <ImageProperties />
           <ColorTokenProperties />
           <FrameProperties />